Extract hierarchy helpers as a toolkit
